feat: Add getAddressFromUserId route and controller
This commit adds a new route and controller function, getAddressFromUserId, to retrieve all addresses associated with a given user ID. The route is accessible via the endpoint "/address/user/:userId". The controller function fetches the addresses from the database based on the user ID and returns them as a response. If no addresses are found for the user ID, an error message is returned.